ref: SfSelect to be native (#1336)
* refactor: wip change sfselect to sfcomponentselect

* fix: unit test fix

* fix: unit tests fix

* refactor: wip create new SfSelect

* refactor: add sfselect native component

* refactor: wip work on select api

* refactor: style select to new design

* fix: fix test

* fix: another test fix

* refactor:  styling select states

* refactor: styling select states remove valid and error message

* refactor: restore valid and errormessage props and styling

* fix: code fixes in component select, after CR
