+++ packages/shared/styles/components/molecules/SfSelect.scss
@@ -1,52 +1,20 @@
 @import "../../helpers";
-.sf-select-option {
-  display: block;
-  box-sizing: border-box;
-  padding: var(--select-option-padding, var(--spacer-sm));
-  @include border(--select-option-border, 1px 0 0 0, solid, var(--c-light));
-  background: var(--select-option-background);
-  color: inherit;
-  min-height: var(
-    --select-height,
-    calc(
-      var(--spacer-base) + var(--spacer-2xs) +
-      var(--select-label-font-size, var(--font-lg))
-    )
-  );
-  @include font(
-    --select-option-font,
-    var(--font-light),
-    var(--font-sm),
-    1.2,
-    var(--font-family-primary)
-  );
-  &--is-active {
-    --select-option-background: var(--c-light);
-  }
-  @include for-desktop {
-    --select-option-padding: var(--spacer-xs);
-    &:hover {
-      --select-option-background: rgba(var(--c-gray-base), 0.1);
-    }
-  }
-}
+
 .sf-select {
   position: relative;
   box-sizing: border-box;
   width: var(--select-width);
-  height: var(--select-height);
-  padding: var(--select-padding, 0 0 calc(var(--font-xs) * 1.2) 0);
-  color: var(--select-color, var(--c-text));
-  cursor: default;
+  height: var(--select-height, 4.25rem);
+  padding: var(--select-padding, var(--spacer-sm) 0 var(--spacer-sm) 0);
+  color: var(--select-color, var(--c-text));      
+  cursor: default;  
   &__label {
     position: absolute;
-    top: var(--select-label-top, 50%);
+    top: var(--select-label-top, 30%);
     left: var(--select-label-left, 0);
+    padding: 0 0 0 4px;
+    height: 0.625rem;
     color: var(--select-label-color, inherit);
-    transform: var(
-      --select-label-transform,
-      translate3d(0, calc(var(--select-label-top, 50%) * -1), 0)
-    );
     transition: var(
       --select-label-transition,
       top 150ms linear,
@@ -54,136 +22,84 @@
     );
     @include font(
       --select-label-font,
-      var(--font-normal),
-      var(--font-lg),
+      var(--font-weight--normal),
+      var(--font-size--base),
       1,
-      var(--font-family-secondary)
+      var(--font-family--secondary)
     );
     &::after {
       content: var(--select-label-required);
       color: var(--select-label-color, var(--c-primary));
     }
   }
-  &__chevron {
-    --chevron-position: absolute;
-    display: var(--chevron-display, block);
-    right: var(--spacer-xs);
-    top: 50%;
-    transform: translate3d(0, -50%, 0);
-  }
-  &__selected {
-    --product-option-font-size: var(--font-lg);
-    --select-option-font-size: var(--font-lg);
-    box-sizing: border-box;
-    display: flex;
-    align-items: var(--select-selected-align-items,center);
-    justify-content: var(--select-selected-justify-content, flex-start);
-    padding: var(
-      --select-selected-padding,
-      var(--spacer-sm) var(--spacer-lg) var(--spacer-xs) 0
-    );
-    margin: var(--select-margin, 0 0 var(--spacer-xs) 0);
-    background: var(--select-background);
-    color: var(--select-color, var(--c-text));
-    @include border(--select-border, 0, solid, transparent);
-  }
   &__dropdown {
-    box-sizing: border-box;
-    position: var(--select-dropdown-position, fixed);
-    top: var(--select-dropdown-top, auto);
-    left: 0;
-    bottom: var(--select-dropdown-bottom, 0);
-    z-index: var(--select-dropdown-z-index, 1);
     width: 100%;
-    background: var(--select-dropdown-background, var(--c-white));
-    box-shadow: var(
-      --select-dropdown-box-shadow,
-      0 4px 11px rgba(var(--c-dark-base), 0.1)
-    );
-  }
-  &__options {
-    margin: 0;
-    padding: 0;
-    list-style-type: none;
+    height: 1.625rem; 
+    margin: 0 0 var(--spacer-2xs) 0;       
     overflow: auto;
-    &::-webkit-scrollbar {
-      width: 0;
-    }
+    box-shadow: none;  
+    color: var(--select-dropdown-color, var(--c-secondary-variant));
+    background: var(--c-white); 
+    outline: none;       
+    @include border(--select-dropdown-border, 0 0 0 0, solid, var(--c-secondary));
   }
-  &__error-message {
-    --select-border-color: var(--c-danger);
+  &__option {           
+    @include background(
+      --select-option-background,
+      var(--c-white)
+    ); 
+    @include font(
+      --select-option-font,
+      var(--font-weight--normal),
+      var(--font-size--base),
+      1,
+      var(--font-family--secondary)
+    ); 
+  }
+  .sf-select__dropdown:active .sf-select__placeholder {
+    display: var(--select-placeholder-display, none);
+  }
+  .sf-select__dropdown:active {
+    --select-label-color: var(--c-text-muted);
+    --select-dropdown-border-color: var(--c-primary);
+    --select-dropdown-color: var(--c-text-muted);
+  }
+  &__label .sf-select__dropdown:active {
+    top: 0;
+  }  
+  &::-ms-expand {
+    display: none;
+  }
+  &__error-message {    
     color: var(--select-error-message-color, var(--c-danger));
-    min-height: var(--select-error-message-height, calc(var(--font-xs) * 1.2));
+    min-height: var(--select-error-message-height, calc(var(--font-size--xs) * 1.2));
     @include font(
       --input-error-message-font,
-      var(--font-medium),
-      var(--font-xs),
+      var(--font-weight--medium),
+      var(--font-size--xs),
       1.2,
-      var(--font-family-secondary)
+      var(--font-family--secondary)
     );
   }
-  &__cancel {
-    --button-background: var(--c-light);
-    --button-color: var(--c-dark-variant);
-    &:hover {
-      --button-background: var(--c-light);
-    }
-  }
   &--underlined {
+    --select-dropdown-border-width: 0 0 1px 0;
     --select-border-width: 0 0 1px 0;
-    --select-border-color: var(--c-light);
+    --select-border-color: var(--c-light);    
   }
   &--is-selected {
     --select-label-top: 0;
-    --select-label-font-size: var(--font-xs);
+    --select-label-font-size: var(--font-size--xs);
+    --select-dropdown-color: var(--c-secondary-variant);
   }
   &--is-invalid {
-    --select-border-color: var(--c-danger);
-  }
-  &--is-active {
-    --chevron-color: var(--c-primary);
-    --select-label-color: var(--c-text-muted);
-    --select-border-color: var(--c-primary);
+    --select-dropdown-border-color: var(--c-danger);
   }
   &--is-disabled {
-    --chevron-color: var(--c-text-disabled);
-    --select-color: var(--c-text-disabled);
-    --select-border-color: var(--c-text-disabled);
+    --select-dropdown-color: var(--c-text-disabled);
+    --select-dropdown-border-color: var(--c-text-disabled);
     --select-label-color: var(--c-text-disabled);
   }
   &--is-required {
     --select-label-required: " *";
-  }
-  &--no-chevron {
-    --chevron-display: none;
-  }
-  &-enter-active {
-    animation: var(--select-animation-enter, ssmobile 150ms);
-  }
-  &-leave-active {
-    animation: var(--select-animation-leave, ssmobile 150ms reverse);
-  }
-  @include for-desktop {
-    --select-dropdown-position: absolute;
-    --select-dropdown-top: 100%;
-    --select-dropdown-bottom: auto;
-    --select-animation-enter: ssdesktop 150ms;
-    --select-animation-leave: ssdesktop 150ms reverse;
-  }
-}
-@keyframes ssdesktop {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-@keyframes ssmobile {
-  0% {
-    transform: translate(0, 100%);
-  }
-  100% {
-    transform: translate(0, 0);
   }
 }
