+++ packages/vue/docs/.vuepress/enhanceApp.js
@@ -28,6 +28,7 @@ import SfBottomModal from "../../src/components/molecules/SfBottomModal/SfBottom
 import SfCallToAction from "../../src/components/molecules/SfCallToAction/SfCallToAction.vue"
 import SfCharacteristic from "../../src/components/molecules/SfCharacteristic/SfCharacteristic.vue"
 import SfCheckbox from "../../src/components/molecules/SfCheckbox/SfCheckbox.vue"
+import SfComponentSelect from "../../src/components/molecules/SfComponentSelect/SfComponentSelect.vue"
 import SfDropdown from "../../src/components/molecules/SfDropdown/SfDropdown.vue"
 import SfFilter from "../../src/components/molecules/SfFilter/SfFilter.vue"
 import SfGallery from "../../src/components/molecules/SfGallery/SfGallery.vue"
@@ -101,6 +102,7 @@ export default ({
   Vue.component("SfCallToAction", SfCallToAction);
   Vue.component("SfCharacteristic", SfCharacteristic);
   Vue.component("SfCheckbox", SfCheckbox);
+  Vue.component("SfComponentSelect", SfComponentSelect);
   Vue.component("SfDropdown", SfDropdown);
   Vue.component("SfFilter", SfFilter);
   Vue.component("SfGallery", SfGallery);
