+++ packages/vue/src/components/molecules/SfComponentSelect/SfComponentSelect.md
@@ -1,4 +1,5 @@
 # component-description
+On click input select field with dropdown
 
 # storybook-iframe-height
-
+10rem
