+++ packages/vue/src/components/molecules/SfComponentSelect/SfComponentSelect.spec.js
@@ -1,0 +1,8 @@
+import { shallowMount } from "@vue/test-utils";
+import SfComponentSelect from "./SfComponentSelect.vue";
+describe("SfComponentSelect.vue", () => {
+  it("renders a component", () => {
+    const component = shallowMount(SfComponentSelect);
+    expect(component.classes("sf-component-select")).toBe(true);
+  });
+});
