+++ packages/vue/src/components/molecules/SfSelect/SfSelect.spec.js
@@ -1,5 +1,6 @@
 import { shallowMount } from "@vue/test-utils";
 import SfSelect from "./SfSelect.vue";
+
 describe("SfSelect.vue", () => {
   it("renders a component", () => {
     const component = shallowMount(SfSelect);
