+++ packages/vue/src/components/molecules/SfSelect/SfSelect.stories.js
@@ -28,7 +28,6 @@ storiesOf("Molecules|Select", module)
           "CSS modifiers",
           {
             "sf-select--underlined": "sf-select--underlined",
-            "sf-select--no-chevron": "sf-select--no-chevron",
           },
           "sf-select--underlined",
           { display: "multi-select" },
@@ -38,8 +37,8 @@ storiesOf("Molecules|Select", module)
       label: {
         default: text("label", "Color", "Props"),
       },
-      size: {
-        default: number("size", 5, {}, "Props"),
+      selected: {
+        default: text("selected", "", "Props"),
       },
       required: {
         default: boolean("required", false, "Props"),
@@ -53,17 +52,8 @@ storiesOf("Molecules|Select", module)
       errorMessage: {
         default: text("errorMessage", "Color", "Props"),
       },
-      persistent: {
-        default: boolean("persistent", false, "Props"),
-      },
-      selected: {
-        default: options(
-          "selected",
-          knobOptionsList,
-          optionsList[0].value,
-          { display: "select" },
-          "Data"
-        ),
+      placeholder: {
+        default: text("placeholder", "", "Props"),
       },
     },
     data() {
@@ -76,13 +66,12 @@ storiesOf("Molecules|Select", module)
         v-model="selected"
         :class="customClass"
         :label="label"
-        :size="size"
         :required="required"
         :valid="valid"
         :disabled="disabled"
         :error-message="errorMessage"
-        :persistent="persistent"
-        style="max-width: 30rem"
+        :placeholder="placeholder"
+        style="max-width: 30rem; margin: 10px;"
       >
         <SfSelectOption v-for="(option, key) in options" :key="key" :value="option.value">
           <SfProductOption :color="option.color" :label="option.label"></SfProductOption>
@@ -107,9 +96,6 @@ storiesOf("Molecules|Select", module)
       label: {
         default: text("label", "Color", "Props"),
       },
-      size: {
-        default: number("size", 5, {}, "Props"),
-      },
       required: {
         default: boolean("required", false, "Props"),
       },
@@ -122,8 +108,8 @@ storiesOf("Molecules|Select", module)
       errorMessage: {
         default: text("errorMessage", "Color", "Props"),
       },
-      persistent: {
-        default: boolean("persistent", false, "Props"),
+      placeholder: {
+        default: text("placeholder", "", "Props"),
       },
     },
     data() {
@@ -137,12 +123,11 @@ storiesOf("Molecules|Select", module)
         v-model="selected"
         :class="customClass"
         :label="label"
-        :size="size"
         :required="required"
         :valid="valid"
-        :disabled="disabled"
-        :error-message="errorMessage" 
-        :persistent="persistent"       
+        :disabled="disabled"  
+        :error-message="errorMessage"
+        :placeholder="placeholder"    
         >
         <SfSelectOption v-for="(option, key) in options" :key="key" :value="option.value">
           <SfProductOption :color="option.color" :label="option.label"></SfProductOption>
@@ -161,7 +146,6 @@ storiesOf("Molecules|Select", module)
           "CSS modifiers",
           {
             "sf-select--underlined": "sf-select--underlined",
-            "sf-select--no-chevron": "sf-select--no-chevron",
           },
           "sf-select--underlined",
           { display: "multi-select" },
@@ -170,9 +154,6 @@ storiesOf("Molecules|Select", module)
       },
       label: {
         default: text("label", "Color", "Props"),
-      },
-      size: {
-        default: number("size", 5, {}, "Props"),
       },
       required: {
         default: boolean("required", false, "Props"),
@@ -186,8 +167,8 @@ storiesOf("Molecules|Select", module)
       errorMessage: {
         default: text("errorMessage", "Color", "Props"),
       },
-      persistent: {
-        default: boolean("persistent", false, "Props"),
+      placeholder: {
+        default: text("placeholder", "", "Props"),
       },
     },
     data() {
@@ -197,87 +178,21 @@ storiesOf("Molecules|Select", module)
       };
     },
     template: `<div style="max-width: 30rem">
-      <SfSelect
+      <SfComponentSelect
         v-model="selected"
         :class="customClass"
         :label="label"
-        :size="size"
         :required="required"
         :valid="valid"
         :disabled="disabled"
         :error-message="errorMessage"
-        :persistent="persistent"
         >
-        <SfSelectOption v-for="(option, key) in options" :key="key" :value="option.value">
+        <SfComponentSelectOption v-for="(option, key) in options" :key="key" :value="option.value">
           <SfProductOption :color="option.color" :label="option.label"></SfProductOption>
-        </SfSelectOption>
+        </SfComponentSelectOption>
         <template #errorMessage>
           CUSTOM ERROR MESSAGE
         </template>
-      </SfSelect>
-    </div>`,
-  }))
-  .add("[slot] cancel", () => ({
-    components: { SfSelect, SfProductOption },
-    props: {
-      customClass: {
-        default: options(
-          "CSS modifiers",
-          {
-            "sf-select--underlined": "sf-select--underlined",
-            "sf-select--no-chevron": "sf-select--no-chevron",
-          },
-          "sf-select--underlined",
-          { display: "multi-select" },
-          "CSS Modifiers"
-        ),
-      },
-      label: {
-        default: text("label", "Color", "Props"),
-      },
-      size: {
-        default: number("size", 5, {}, "Props"),
-      },
-      required: {
-        default: boolean("required", false, "Props"),
-      },
-      valid: {
-        default: boolean("valid", false, "Props"),
-      },
-      disabled: {
-        default: boolean("disabled", false, "Props"),
-      },
-      errorMessage: {
-        default: text("errorMessage", "Color", "Props"),
-      },
-      persistent: {
-        default: boolean("persistent", false, "Props"),
-      },
-    },
-    data() {
-      return {
-        selected: "",
-        options: optionsList,
-      };
-    },
-    template: `<div style="max-width: 30rem">
-      <SfSelect
-        v-model="selected"
-        :class="customClass"
-        :label="label"
-        :size="size"
-        :required="required"
-        :valid="valid"
-        :disabled="disabled"
-        :error-message="errorMessage"
-        :persistent="persistent"
-        >
-        <SfSelectOption v-for="(option, key) in options" :key="key" :value="option.value">
-          <SfProductOption :color="option.color" :label="option.label"></SfProductOption>
-        </SfSelectOption>
-        <template #cancel>
-          CUSTOM CANCEL BUTTON
-        </template>
-      </SfSelect>
+      </SfComponentSelect>
     </div>`,
   }));
