+++ packages/vue/src/components/molecules/SfSelect/SfSelect.stories.scss
@@ -1,0 +1,1 @@
+@import "~@storefront-ui/shared/styles/helpers";
