+++ packages/vue/src/components/molecules/SfSelect/SfSelect.vue
@@ -1,70 +1,32 @@
 <template>
   <div
-    v-click-outside="checkPersistence"
-    :aria-expanded="open.toString()"
-    :aria-owns="'lbox_' + _uid"
-    :aria-label="label"
-    role="listbox"
+    class="sf-select"
     :class="{
-      'sf-select--is-active': isActive,
-      'sf-select--is-selected': isSelected,
+      'sf-select--is-selected': selected || placeholder,
       'sf-select--is-required': required,
       'sf-select--is-disabled': disabled,
       'sf-select--is-invalid': !valid,
     }"
-    class="sf-select"
-    @click="toggle($event)"
-    @keyup.esc="closeHandler"
-    @keyup.space="openHandler"
-    @keyup.up="move(-1)"
-    @keyup.down="move(1)"
-    @keyup.enter="enter($event)"
   >
-    <div style="position: relative;">
-      <div
-        ref="sfSelect"
-        v-focus
-        tabindex="0"
-        class="sf-select__selected sf-select-option"
-        v-on="$listeners"
-        v-html="html"
-      ></div>
-      <slot name="label">
-        <div v-if="label" class="sf-select__label">
-          {{ label }}
-        </div>
+    <label :for="label" class="sf-select__label">
+      <slot name="label" :label="label">
+        {{ label }}
       </slot>
-      <slot name="icon">
-        <SfChevron class="sf-select__chevron" />
-      </slot>
-      <SfOverlay
-        ref="overlay"
-        :visible="open"
-        class="sf-select__overlay mobile-only"
-      />
-      <transition name="sf-select">
-        <div v-show="open" class="sf-select__dropdown">
-          <!--  sf-select__option -->
-          <ul
-            :aria-expanded="open.toString()"
-            :style="{ maxHeight }"
-            class="sf-select__options"
-          >
-            <slot />
-          </ul>
-          <slot name="cancel">
-            <SfButton
-              ref="cancel"
-              class="sf-select__cancel sf-button--full-width mobile-only"
-              @click="closeHandler"
-            >
-              Cancel
-            </SfButton>
-          </slot>
-        </div>
-      </transition>
-    </div>
-    <div class="sf-select__error-message">
+    </label>
+    <select :id="label" v-model="selected" v-focus class="sf-select__dropdown">
+      <!-- empty option by default, may be used as placeholder -->
+      <option
+        class="sf-select__placeholder sf-select__option"
+        disabled
+        selected
+        value
+        ><slot name="placeholder" :placeholder="placeholder" />{{
+          placeholder
+        }}</option
+      >
+      <slot />
+    </select>
+    <div class="sf-component-select__error-message">
       <transition name="sf-fade">
         <!-- @slot Custom error message of form select -->
         <slot v-if="!valid" name="error-message" v-bind="{ errorMessage }">
@@ -75,26 +37,13 @@
   </div>
 </template>
 <script>
-import SfSelectOption from "./_internal/SfSelectOption.vue";
-import SfChevron from "../../atoms/SfChevron/SfChevron.vue";
-import SfButton from "../../atoms/SfButton/SfButton.vue";
-import SfOverlay from "../../atoms/SfOverlay/SfOverlay.vue";
 import { focus } from "../../../utilities/directives";
-import { clickOutside } from "../../../utilities/directives";
+import SfSelectOption from "./_internal/SfSelectOption.vue";
 import Vue from "vue";
 Vue.component("SfSelectOption", SfSelectOption);
 export default {
   name: "SfSelect",
-  directives: { focus, clickOutside },
-  components: {
-    SfButton,
-    SfChevron,
-    SfOverlay,
-  },
-  model: {
-    prop: "selected",
-    event: "change",
-  },
+  directives: { focus },
   props: {
     /**
      * Select field label
@@ -109,13 +58,6 @@ export default {
     selected: {
       type: [String, Number, Object],
       default: "",
-    },
-    /**
-     * Dropdown list size
-     */
-    size: {
-      type: Number,
-      default: 5,
     },
     /**
      * Required attribute
@@ -146,136 +88,16 @@ export default {
       default: "This field is not correct.",
     },
     /**
-     * If true clicking outside will not dismiss the select
+     * Adds placeholder
      */
-    persistent: {
-      type: Boolean,
-      default: false,
+    placeholder: {
+      type: String,
+      default: "",
     },
-  },
-  data() {
-    return {
-      open: false,
-      options: [],
-      indexes: {},
-      optionHeight: 0,
-      focusedOption: "",
-    };
   },
   computed: {
-    index: {
-      get() {
-        const stringified = this.indexes[JSON.stringify(this.selected)];
-        if (typeof stringified === "undefined") {
-          return -1;
-        }
-        return stringified;
-      },
-      set(index) {
-        this.focusedOption = this.options[index].value;
-        this.$emit("change", this.options[index].value);
-      },
-    },
-    html() {
-      if (this.index < 0) return;
-      return this.options[this.index].html;
-    },
-    maxHeight() {
-      if (!this.options.length) return;
-      return `${this.optionHeight * this.options.length}px`;
-    },
-    isActive() {
-      return this.open;
-    },
     isSelected() {
       return this.selected;
-    },
-  },
-  watch: {
-    open: {
-      immediate: true,
-      handler: function (visible) {
-        if (visible) {
-          this.$nextTick(() => {
-            if (this.$slots.default) {
-              this.optionHeight = this.$slots.default[0].elm.offsetHeight;
-            }
-          });
-        }
-      },
-    },
-  },
-  created: function () {},
-  mounted: function () {
-    this.addOptionsAndIndexes();
-  },
-  updated() {
-    if (this.$slots.default) {
-      if (this.$slots.default.length > this.options.length) {
-        this.addOptionsAndIndexes();
-      }
-    }
-  },
-  beforeDestroy: function () {
-    this.$off("update", this.update);
-  },
-  methods: {
-    update(index) {
-      this.index = index;
-    },
-    addOptionsAndIndexes() {
-      const options = [];
-      const indexes = {};
-      if (!this.$slots.default) return;
-      this.$on("update", this.update);
-      this.$slots.default.forEach(({ tag, componentOptions, elm }, index) => {
-        if (!tag) return;
-        options.push({
-          ...componentOptions.propsData,
-          html: elm.innerHTML,
-        });
-        indexes[JSON.stringify(componentOptions.propsData.value)] = index;
-      });
-      this.options = options;
-      this.indexes = indexes;
-    },
-    move(payload) {
-      const optionsLength = this.options.length;
-      let index = this.index;
-      index += payload;
-      if (index < 0) index = 0;
-      if (index >= optionsLength) index = optionsLength - 1;
-      this.index = index;
-      this.$refs.sfSelect.blur();
-      document.getElementById(this.focusedOption).focus();
-    },
-    enter() {
-      this.toggle();
-    },
-    toggle(event) {
-      if (
-        (this.$refs.cancel &&
-          event &&
-          event.target.contains(this.$refs.cancel.$el)) ||
-        (this.$refs.overlay &&
-          event &&
-          this.persistent &&
-          event.target.contains(this.$refs.overlay.$el)) ||
-        this.disabled
-      )
-        return;
-      this.open = !this.open;
-    },
-    checkPersistence() {
-      if (!this.persistent) {
-        this.closeHandler();
-      }
-    },
-    openHandler() {
-      this.open = true;
-    },
-    closeHandler() {
-      this.open = false;
     },
   },
 };
