+++ packages/vue/src/components/molecules/SfSelect/_internal/SfSelectOption.spec.js
@@ -1,8 +1,9 @@
 import { shallowMount } from "@vue/test-utils";
 import SfSelectOption from "./SfSelectOption.vue";
+
 describe("SfSelectOption.vue", () => {
   it("renders a component", () => {
     const component = shallowMount(SfSelectOption);
-    expect(component.classes("sf-select-option")).toBe(true);
+    expect(component.classes("sf-select__option")).toBe(true);
   });
 });
