+++ packages/vue/src/components/molecules/SfSelect/_internal/SfSelectOption.vue
@@ -1,23 +1,12 @@
 <template>
-  <li
-    :id="value"
-    role="option"
-    class="sf-select-option"
-    :class="{ 'sf-select-option--is-active': selected === value }"
-    tabindex="0"
-    :aria-selected="selected === value ? 'true' : 'false'"
-    @click="clicked"
-    @keyup.space.enter="clicked"
-  >
+  <option :value="value" class="sf-select__option" tabindex="0">
     <!-- @slot -->
     <slot />
-  </li>
+  </option>
 </template>
 <script>
-import { focus } from "../../../../utilities/directives";
 export default {
   name: "SfSelectOption",
-  directives: { focus },
   props: {
     value: {
       type: [String, Number, Object],
@@ -27,14 +16,6 @@ export default {
   computed: {
     selected() {
       return this.$parent.selected;
-    },
-    indexes() {
-      return this.$parent.indexes;
-    },
-  },
-  methods: {
-    clicked() {
-      this.$parent.$emit("update", this.indexes[JSON.stringify(this.value)]);
     },
   },
 };
